+++ dirty_cat/test/test_utils.py
@@ -7,11 +7,11 @@ def test_lrudict():
     dict_ = LRUDict(10)
 
     for x in range(15):
-        dict_[x] = 'filled'+str(x)
+        dict_[x] = f'filled {x}'
 
     for x in range(5, 15):
         assert x in dict_
-        assert dict_[x] == 'filled'+str(x)
+        assert dict_[x] == f'filled {x}'
 
     for x in range(5):
         assert (x not in dict_)